+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Fix Text metric publication when several differents labels are used
+    - Fix JSON schemas
 
 ### 4.27.1